Verificação de erros

+++ src/pages/repositories/index.js
@@ -1,16 +1,23 @@
 import React, { useEffect, useState } from 'react';
+import { useHistory } from 'react-router-dom';
 import * as S from './styled';
 
 export default function Repositories() {
 
     const [repositories, setRepositories] = useState([]);
+    const history = useHistory();
 
     useEffect(() => {
 
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if (repositoriesName != null) {
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        } else {
+            history.push('/');
+        }
+
 
     }, []);
 
@@ -21,10 +28,10 @@ export default function Repositories() {
             <S.Title>Repositórios</S.Title>
             <S.List>
                 {repositories.map(repository => {
-                    return (<S.listItem>{repository}</S.listItem>)
-
+                    return (<S.listItem>Repositório: {repository}</S.listItem>)
                 })}
             </S.List>
+            <S.linkHome to='/'>Voltar</S.linkHome>
         </S.Container>
 
     )
